requirements.txt
Scrapy==2.3.0
psycopg2-binary==2.8.6
